server/tasks.py
import time
import re
import boto3
import os
import subprocess
import datetime
import zipfile
import jinja2
import random
import json

from backports import tempfile

from server import celery_app, app, db, socketio
from server.models import Bot, Team, Game
from server.helpers import get_s3_object

from sqlalchemy.orm import raiseload
