requirements.txt
alembic==1.0.5
amqp==2.1.4
Babel==2.6.0
backports-abc==0.5
backports.tempfile==1.0rc1
backports.weakref==1.0rc1
billiard==3.5.0.2
boto3==1.9.71
botocore==1.12.71
celery==4.2.0
certifi==2018.11.29
chardet==3.0.4
click==6.6
coolname==1.1.0
docutils==0.14
enum34==1.1.6
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SocketIO==3.2.1
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
flower==0.9.2
futures==3.0.5
gunicorn==19.9.0
humanize==0.5.1
idna==2.8
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.0
kombu==4.2.1
Mako==1.0.7
MarkupSafe==0.23
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-editor==1.0.3
python-engineio==3.3.0
python-socketio==3.1.1
pytz==2018.7
redis==2.10.6
requests==2.21.0
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.1.4
tornado==5.1.1
urllib3==1.24.1
vine==1.1.3
Werkzeug==0.14.1
